feat(ui): rename main app components
